ms_chat/requirements.txt
fastapi==0.68.0
sqlalchemy==1.4.22
sqlmodel==0.0.7
kafka-python==2.0.2
websockets==10.0